+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * Bug fixes
 * Performance improvements
 * Better support for int type
+* New operation: diff2'', original value, first derivative, and second derivative of a scalar-to-scalar function
 -->
 
 #### 0.5.0 - October 2, 2014